requirements.txt
pytelegrambotapi
dotenv
GoogleBard==1.4.0
